+++ src/components/Home/HomePlaylist.tsx
@@ -1,17 +1,32 @@
-import { useEffect, useRef, useState } from "react";
+import { ChangeEvent, useEffect, useRef, useState } from "react";
 import DBType from "../../type/DBType";
 import PagesType from "../../type/PagesType";
 import MessageCom from "../MessageCom";
 import PlaylistCard from "../Card/PlaylistCard";
+import PlaylistCardType from "../../type/PlaylistCardType";
+import JsonSong from "../../type/Songs";
 
 interface pageProps {
     pageProps: () => PagesType;
     dbProp: DBType;
 }
 
+type jsonImport = {
+    playlist: PlaylistCardType;
+    songs: JsonSong[];
+}
+
 function HomePlaylist({ pageProps, dbProp }: pageProps) {
     const [component, setComponent] = useState<JSX.Element[]>([]);
     const inputRef = useRef<HTMLInputElement>(null);
+    const [dataImport, setDataImport] = useState<jsonImport | null>(null);
+    const [selectedOption, setSelectedOption] = useState<string>('create');
+    const [allPlaylist, setAllPlaylist] = useState<PlaylistCardType[]>([]);
+    const [showBtnClose, setShowBtnClose] = useState<boolean>(true);
+
+    const handleOptionChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+        setSelectedOption(event.target.value);
+    };
 
     useEffect(() => {
         setComponent([<MessageCom key={"512"} msg="Pesquisando playlist, aguarde...." />])
@@ -20,6 +35,7 @@ function HomePlaylist({ pageProps, dbProp }: pageProps) {
 
     async function getAllPlaylist() {
         const result = await dbProp.getAllPlaylist();
+        setAllPlaylist(result);
 
         const comp = [] as JSX.Element[];
         result.forEach((v) => {
@@ -31,9 +47,140 @@ function HomePlaylist({ pageProps, dbProp }: pageProps) {
     const createPlaylist = () => {
         if (inputRef.current) {
             const name = inputRef.current.value;
-            dbProp.createPlaylist(name);
-            getAllPlaylist();
+            createPlaylistByName(name);
         }
+    }
+
+    const createPlaylistByName = async (name: string) => {
+        dbProp.createPlaylist(name);
+        await getAllPlaylist();
+    }
+
+    const isValidJsonImport = (obj: any): obj is jsonImport => {
+        return obj && obj.playlist && Array.isArray(obj.songs);
+    };
+
+
+    const onChangeImportJson = (e: ChangeEvent<HTMLInputElement>) => {
+        const file = e.target.files?.[0];
+        if (file) {
+            const reader = new FileReader();
+            reader.onload = (event) => {
+                const result = event.target?.result;
+                if (typeof result === 'string') {
+                    const jsonData: jsonImport = JSON.parse(result);
+                    if (isValidJsonImport(jsonData))
+                        setDataImport(jsonData);
+                    else
+                        setDataImport(null);
+
+                } else
+                    setDataImport(null);
+            };
+            reader.readAsText(file);
+        } else
+            setDataImport(null);
+    }
+
+    const importSongsFromJson = () => {
+        if (dataImport) {
+            console.log(selectedOption, dataImport);
+            setShowBtnClose(false);
+            switch (selectedOption) {
+                case "create":
+                    createAnyaway();
+                    break;
+                case "mix":
+                    mixPlaylist();
+                    break;
+                case "replace":
+                    replacePlaylist();
+                    break;
+                default:
+                    createAnyaway();
+                    break;
+            }
+        }
+    }
+
+    const createAnyaway = async () => {
+        let name = dataImport?.playlist.title as string;
+
+        for (let i = 0; i < allPlaylist.length; i++) {
+            if (allPlaylist[i].title === name) {
+                const d = new Date();
+                name += " " + d.toLocaleDateString() + " " + d.toLocaleTimeString();
+                break;
+            }
+        }
+
+        await createPlaylistByName(name);
+        const result = await dbProp.getAllPlaylist();
+        const id = result[result.length - 1].id;
+        const songs = dataImport?.songs as JsonSong[]
+
+        await dbProp.saveSongList(songs);
+        for (let i = 0; i < songs.length; i++)
+            await dbProp.addSongInPlaylist(id, songs[i].annSongId);
+
+        await getAllPlaylist();
+        setShowBtnClose(true);
+    }
+
+    const mixPlaylist = async () => {
+        let id = -1;
+
+        for (let i = 0; i < allPlaylist.length; i++) {
+            if (allPlaylist[i].title === dataImport?.playlist.title) {
+                id = allPlaylist[i].id;
+                break;
+            }
+        }
+
+        const songs = dataImport?.songs as JsonSong[]
+        await dbProp.saveSongList(songs);
+
+        if (id != -1)
+            for (let i = 0; i < songs.length; i++)
+                await dbProp.addSongInPlaylist(id, songs[i].annSongId);
+        else {
+            await createPlaylistByName(dataImport?.playlist.title as string);
+            const result = await dbProp.getAllPlaylist();
+            const id_playlist = result[result.length - 1].id;
+
+            for (let i = 0; i < songs.length; i++)
+                await dbProp.addSongInPlaylist(id_playlist, songs[i].annSongId);
+        }
+
+        await getAllPlaylist();
+        setShowBtnClose(true);
+    }
+
+    const replacePlaylist = async () => {
+        let id = -1;
+
+        for (let i = 0; i < allPlaylist.length; i++) {
+            if (allPlaylist[i].title === dataImport?.playlist.title) {
+                id = allPlaylist[i].id;
+                break;
+            }
+        }
+
+        const songs = dataImport?.songs as JsonSong[]
+        await dbProp.saveSongList(songs);
+
+        if (id != -1)
+            await dbProp.deletePlaylist(id);
+
+        await createPlaylistByName(dataImport?.playlist.title as string);
+        const result = await dbProp.getAllPlaylist();
+        const id_playlist = result[result.length - 1].id;
+
+        for (let i = 0; i < songs.length; i++)
+            await dbProp.addSongInPlaylist(id_playlist, songs[i].annSongId);
+
+        await getAllPlaylist();
+        setShowBtnClose(true);
     }
 
     return (
@@ -47,7 +194,7 @@ function HomePlaylist({ pageProps, dbProp }: pageProps) {
                         </div>
                         <div className="w-100 d-flex justify-content-center">
                             <a href="#" className="btn btn-outline-success mt-2" data-bs-toggle="modal" data-bs-target="#playlistModal">Create Playlist</a>
-                            <a href="#" className="btn btn-outline-secondary disabled mt-2 ms-1" data-bs-toggle="modal" data-bs-target="#playlistModal">Import playlist</a>
+                            <a href="#" className="btn btn-outline-secondary mt-2 ms-1" data-bs-toggle="modal" data-bs-target="#importModal">Import playlist</a>
                         </div>
                     </div>
                 </div>
@@ -69,6 +216,60 @@ function HomePlaylist({ pageProps, dbProp }: pageProps) {
                                 <div className="modal-footer">
                                     <button type="button" className="btn btn-outline-success" onClick={createPlaylist} data-bs-dismiss="modal">Create</button>
                                     <button type="button" className="btn btn-outline-secondary" data-bs-dismiss="modal">Close</button>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                    <div className="modal fade" id="importModal" tabIndex={-1} aria-labelledby="importModalTitle" aria-hidden="true" data-bs-backdrop="static" data-bs-keyboard="false" >
+                        <div className="modal-dialog modal-lg">
+                            <div className="modal-content ">
+                                <div className="modal-header">
+                                    <h1 className="modal-title fs-5" id="importModalTitle">Import Playlist</h1>
+                                    {showBtnClose &&
+                                        <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                                    }
+                                </div>
+                                <div className="modal-body">
+                                    <div className="mb-3">
+                                        <input className="form-control" type="file" accept=".json" onChange={onChangeImportJson} id="formFile" />
+                                    </div>
+                                    <div>
+                                        <h5>In case of playlist name conflict:</h5>
+                                    </div>
+                                    <div>
+                                        <div className="form-check form-check-inline">
+                                            <input className="form-check-input" onChange={handleOptionChange} type="radio" name="importOptionRadio" id="inlineRadio3" value="create" defaultChecked />
+                                            <label className="form-check-label" htmlFor="inlineRadio3">Create under another name</label>
+                                        </div>
+                                        <div className="form-check form-check-inline">
+                                            <input className="form-check-input" onChange={handleOptionChange} type="radio" name="importOptionRadio" id="inlineRadio2" value="mix" />
+                                            <label className="form-check-label" htmlFor="inlineRadio2">Mix</label>
+                                        </div>
+                                        <div className="form-check form-check-inline">
+                                            <input className="form-check-input" onChange={handleOptionChange} type="radio" name="importOptionRadio" id="inlineRadio1" value="replace" />
+                                            <label className="form-check-label" htmlFor="inlineRadio1">Replace everything</label>
+                                        </div>
+                                    </div>
+                                    {dataImport &&
+                                        <div>
+                                            <hr />
+                                            <h5>{dataImport.playlist.title}</h5>
+                                            <p>{dataImport.songs.length} songs</p>
+                                        </div>
+                                    }
+                                    {!showBtnClose &&
+                                        <div className="spinner-border" role="status">
+                                            <span className="visually-hidden">Loading...</span>
+                                        </div>
+                                    }
+                                </div>
+                                <div className="modal-footer">
+                                    {(dataImport && showBtnClose) &&
+                                        <button type="button" onClick={importSongsFromJson} className="btn btn-outline-success">Import</button>
+                                    }
+                                    {showBtnClose &&
+                                        <button type="button" className="btn btn-outline-secondary" data-bs-dismiss="modal">Close</button>
+                                    }
                                 </div>
                             </div>
                         </div>