+++ changelog.md
@@ -43,6 +43,10 @@
 * Importing json into the system coming soon
 
 
+## (03/03/2024)
+* Importing playlist json 
+
+
 <hr>
 
 Anime Song Player © 2023 by 43D is licensed under CC BY-NC-SA 4.0. To view a copy of this license, visit http://creativecommons.org/licenses/by-nc-sa/4.0/